requirements.txt
aiobotocore==2.13.2
aiohappyeyeballs==2.3.7
aiohttp==3.10.4
aioitertools==0.11.0
aiosignal==1.3.1
anyio==4.0.0
appdirs==1.4.4
arf==2.6.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arm_pyart==1.18.7
arrow==1.3.0
asttokens==2.4.0
async-lru==2.0.4
attrs==23.1.0
Babel==2.13.0
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.1.0
blosc2==2.4.0
botocore==1.34.131
certifi==2023.7.22
cffi==1.16.0
cftime==1.6.4
charset-normalizer==3.3.0
click==8.1.7
cloudpickle==3.0.0
cmweather==0.3.2
comm==0.1.4
contourpy==1.1.1
cramjam==2.7.0
cycler==0.12.1
dask==2024.8.1
debugpy==1.8.0
decide-analysis==2.2.0
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.7
ewave==1.0.7
executing==2.0.0
fastjsonschema==2.18.1
fastparquet==2023.10.1
fonttools==4.43.1
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2024.6.1
Gammatone==1.0.1
h11==0.14.0
h5netcdf==1.3.0
h5py==3.10.0
httpcore==0.18.0
httpx==0.25.0
idna==3.4
imageio==2.34.1
importlib_metadata==8.2.0
ipykernel==6.25.2
ipython==8.16.1
ipython-genutils==0.2.0
ipywidgets==8.1.1
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.8.0
jupyter-lsp==2.2.0
jupyter_client==8.4.0
jupyter_core==5.4.0
jupyter_server==2.8.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.7
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
jupyterlab_server==2.25.0
kiwisolver==1.4.5
lat-lon-parser==1.3.0
lazy_loader==0.4
libtfr==2.1.7
locket==1.0.0
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
mda-xdrlib==0.2.0
melizalab-gammatone==0.1.1
melizalab-tools==2024.1.29
mistune==3.0.2
mpltern==1.0.4
msgpack==1.0.7
multidict==6.0.4
nbclient==0.8.0
nbconvert==7.9.2
nbformat==5.9.2
ndindex==1.7
nest-asyncio==1.5.8
netCDF4==1.7.1.post2
networkx==3.3
neurobank==0.10.8
notebook==7.0.6
notebook_shim==0.2.3
numexpr==2.8.8
numpy==1.26.0
open-radar-data==0.3
overrides==7.4.0
packaging==23.2
pandarallel==1.6.5
pandas==2.2.2
pandoc==2.3
pandocfilters==1.5.0
parse==1.20.0
parso==0.8.3
partd==1.4.2
patsy==0.5.6
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
platformdirs==3.11.0
plumbum==1.8.2
ply==3.11
pooch==1.8.2
melizalab-gammatone==0.1.1
prometheus-client==0.17.1
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==14.0.2
pycparser==2.21
Pygments==2.16.1
pyparsing==3.1.1
pyproj==3.6.1
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
qtconsole==5.4.4
QtPy==2.4.0
referencing==0.30.2
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.10.6
s3fs==2024.6.1
samplerate==0.1.0
scikit-image==0.23.2
scikit-learn==1.3.2
scipy==1.11.3
seaborn==0.13.2
Send2Trash==1.8.2
single-source==0.3.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
stack-data==0.6.3
statsmodels==0.14.1
tables==3.9.2
terminado==0.17.1
threadpoolctl==3.2.0
tifffile==2024.5.22
tinycss2==1.2.1
toelis==2.1.1
toolz==0.12.1
tornado==6.3.3
traitlets==5.11.2
types-python-dateutil==2.8.19.14
tzdata==2023.3
uri-template==1.3.0
urllib3==2.0.7
wcwidth==0.2.8
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
widgetsnbextension==4.0.9
wrapt==1.16.0
xarray==2024.7.0
xarray-datatree==0.0.14
xmltodict==0.13.0
xradar==0.6.3
yarl==1.9.4
zipp==3.20.0
